webpack applied

+++ src/controllers/videoController.js
@@ -1,3 +1,4 @@
+import User from "../models/User";
 import Video from "../models/Video";
 
 export const home = async (req, res) => {
@@ -23,17 +24,21 @@ export const search = async (req, res) => {
 
 export const see = async (req, res) => {
   const { id } = req.params;
-  const video = await Video.findById(id);
-  console.log(video);
+  const video = await Video.findById(id).populate("owner");
   if (!video) return res.render("404", { pageTitle: "Video Not Found" });
   else return res.render("watch", { pageTitle: video.title, video });
 };
 export const getEdit = async (req, res) => {
-  const { id } = req.params;
+  const {
+    params: { id },
+    session: {
+      user: { _id },
+    },
+  } = req;
   const video = await Video.findById(id);
   if (!video) return res.render("404", { pageTitle: "Video Not Found" });
-  else
-    return res.render("editVideo", { pageTitle: `Edit ${video.title}`, video });
+  if (String(video.owner) !== String(_id)) return res.status(403).redirect("/");
+  return res.render("editVideo", { pageTitle: `Edit ${video.title}`, video });
 };
 export const postEdit = async (req, res) => {
   const { id } = req.params;
@@ -52,15 +57,24 @@ export const getUpload = (req, res) => {
   return res.render("upload", { pageTitle: "Upload Video" });
 };
 export const postUpload = async (req, res) => {
-  const { path: fileUrl } = req.file;
-  const { title, description, hashtags } = req.body;
+  const {
+    file: { path: fileUrl },
+    body: { title, description, hashtags },
+    session: {
+      user: { _id },
+    },
+  } = req;
   try {
-    await Video.create({
+    const video = await Video.create({
       title,
       fileUrl,
+      owner: _id,
       description,
       hashtags: Video.formatHashtags(hashtags),
     });
+    const user = await User.findById(video.owner);
+    user.videos.push(video._id);
+    user.save();
     return res.redirect("/");
   } catch (error) {
     return res.status(404).render("upload", {
@@ -70,7 +84,15 @@ export const postUpload = async (req, res) => {
   }
 };
 export const remove = async (req, res) => {
-  const { id } = req.params;
+  const {
+    params: { id },
+    session: {
+      user: { _id },
+    },
+  } = req;
+  const video = await Video.findById(id);
+  if (!video) return res.render("404", { pageTitle: "Video Not Found" });
+  if (String(video.owner) !== String(_id)) return res.status(403).redirect("/");
   await Video.findByIdAndDelete(id);
   return res.redirect("/");
 };
